requirements.txt
# Python dependencies for Fatih Agent
# Core LLM APIs
anthropic>=0.18.0
openai>=1.12.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# CLI and UI
click>=8.1.0
rich>=13.0.0

# HTTP and networking
requests>=2.31.0
httpx>=0.26.0

# Docker SDK
docker>=7.0.0

# Retry and resilience
tenacity>=8.2.0

# Parsing and data processing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Browser automation for JavaScript rendering
playwright>=1.40.0
